Moved hdf functions to hdf_utils
